refactor(fmtlog)!: move fmtlog from /fmtlog to /plugins/fmtlog
The fmtlog plugin lived at the top of the repo, alongside `plugins/`,
while every other LogLayer plugin (`redact`, `sampling`, `datadogtrace`,
`oteltrace`) lived under `plugins/`. The docs grouped fmtlog under
"Available Plugins" and linked to `/plugins/fmtlog`, but the actual
import path was bare `go.loglayer.dev/fmtlog` — so users saw the page
in the Plugins sidebar, copied the install line, and ended up with an
inconsistent import path relative to neighbours.

This commit aligns code layout and import path with the rest of the
plugin family. The package stays bundled in the main module (no separate
`go.mod`), matching `plugins/redact` and `plugins/sampling`.

Behaviour is unchanged: same package name (`fmtlog`), same
`fmtlog.New()` constructor, same `MessageHook` semantics. Only the
import path changes.

BREAKING CHANGE: the fmtlog plugin's import path changes from
`go.loglayer.dev/fmtlog` to `go.loglayer.dev/plugins/fmtlog`.
Update imports accordingly.
